Preserve the rebuilt v3.3.0 schema and RPC refactors while applying the final DynamicMPT ImmutableFlags semantics. Refresh the generated ledger catalog and pin the resolved schema order to the rippled v3.3.0 oracle.

+++ internal/ledger/service/ledger_file_test.go
@@ -101,7 +101,7 @@ func TestLoadLedgerJSONWrappersAndDefaults(t *testing.T) {
 			if !gotHeader.CloseTime.Equal(test.wantCloseTime) {
 				t.Errorf("CloseTime = %v, want %v", gotHeader.CloseTime, test.wantCloseTime)
 			}
-			if gotHeader.CloseTimeResolution != test.wantResolution {
+			if uint32(gotHeader.CloseTimeResolution) != test.wantResolution {
 				t.Errorf("CloseTimeResolution = %d, want %d", gotHeader.CloseTimeResolution, test.wantResolution)
 			}
 			if gotHeader.Drops != test.wantDrops {
